Add SignIn and SignOut
