+++ src/App.js
@@ -1,23 +1,22 @@
-import { db, googleAuthProvider, firebase } from "./config/firebase";
+//import { db } from "./config/firebase";
 import Button from "./components/Button";
+import { signInWithGoogle, signOut  } from './login';
 
 function App() {
-  const signInWithGoogle = async () => {
-    firebase.auth().useDeviceLanguage();
-
-    try{
-      await firebase.auth().signInWithPopup(googleAuthProvider);
-
-    } catch (e) {
-      console.error (e.message);
-    }
-  }
+ 
   return (
     <div>
-          <Button onClick = {signInWithGoogle} > Sing in with Google </Button>
-    </div>
-  )
-};
+      {
+        user ? (
+          <>
+          <Button onClick={ signOut } > Sing in with Google</Button>
+           <p> Bienvenidos al CHAT!</p>
+           </>
+        ) : <Button onClick = {signInWithGoogle} > Sing in with Google </Button>
+        }
+        </div>
+  );
+}
 
 export default App;
