+++ src/login/index.js
@@ -1,0 +1,8 @@
+
+import signInWithGoogle from './signInWithGoogle';
+import signOut from './signOut.js';
+
+export {
+    signInWithGoogle,
+    signOut,
+}
